+++ src/main/java/de/dennisthegamer/shulkersort/config/OverflowMode.java
@@ -1,6 +1,0 @@
-package de.dennisthegamer.shulkersort.config;
-
-public enum OverflowMode {
-    FILL,
-    DOMINANT
-}