+++ src/main/java/de/dennisthegamer/shulkersorter/config/OverflowMode.java
@@ -1,0 +1,6 @@
+package de.dennisthegamer.shulkersorter.config;
+
+public enum OverflowMode {
+    FILL,
+    DOMINANT
+}